hdd_menu: add again check_kernel_fs

+++ src/gui/hdd_menu.cpp
@@ -145,6 +145,14 @@ int CHDDMenuHandler::filterDevName(const char * name)
 	return 0;
 }
 
+static std::string readlink(const char *path)
+{
+	char link[PATH_MAX + 1];
+	if (realpath(path, link))
+		return std::string(link);
+	return "";
+}
+
 bool CHDDMenuHandler::is_mounted(const char *dev)
 {
 	bool res = false;
@@ -154,14 +162,25 @@ bool CHDDMenuHandler::is_mounted(const char *dev)
 	else
 		snprintf(devpath, sizeof(devpath), "/dev/%s", dev);
 
-	int devpathlen = strlen(devpath);
 	char buffer[255];
+	std::string realdev = readlink(devpath);
+	realdev = trim(realdev);
 	FILE *f = fopen("/proc/mounts", "r");
 	if(f) {
-		while (!res && fgets(buffer, sizeof(buffer), f))
-			if (!strncmp(buffer, devpath, devpathlen)
-					&& (buffer[devpathlen] == ' ' || buffer[devpathlen] == '\t'))
+		while (!res && fgets(buffer, sizeof(buffer), f)) {
+			if (buffer[0] != '/')
+				continue; /* only "real" devices are interesting */
+			char *p = strchr(buffer, ' ');
+			if (p)
+				*p = 0; /* terminate at first space, kernel-user ABI is fixed */
+			if (!strcmp(buffer, devpath)) /* default '/dev/sda1' mount */
 				res = true;
+			else {	/* now the case of '/dev/disk/by-label/myharddrive' mounts */
+				std::string realmount = readlink(buffer);
+				if (realdev == trim(realmount))
+					res = true;
+			}
+		}
 		fclose(f);
 	}
 	printf("CHDDMenuHandler::is_mounted: dev [%s] is %s\n", devpath, res ? "mounted" : "not mounted");
@@ -283,9 +302,36 @@ std::string CHDDMenuHandler::getFmtType(std::string name, std::string part)
 	return ret;
 }
 
+void CHDDMenuHandler::check_kernel_fs()
+{
+	char line[128]; /* /proc/filesystems lines are shorter */
+	kernel_fs_list.clear();
+	FILE *f = fopen("/proc/filesystems", "r");
+	if (! f) {
+		fprintf(stderr, "CHDDMenuHandler::%s: opening /proc/filesystems failed: %m\n", __func__);
+		return;
+	}
+	while (fgets(line, sizeof(line), f)) {
+		size_t l = strlen(line);
+		if (l > 0)
+			line[l - 1] = 0; /* remove \n */
+		char *tab = strchr(line, '\t');
+		if (! tab)	/* should not happen in any kernel I have seen */
+			continue;
+		tab++;
+		kernel_fs_list.insert(std::string(tab));
+	}
+	fclose(f);
+}
+
 void CHDDMenuHandler::check_dev_tools()
 {
 	for (unsigned i = 0; i < FS_MAX; i++) {
+		if (kernel_fs_list.find(devtools[i].fmt) == kernel_fs_list.end()) {
+			printf("%s: filesystem '%s' not supported by kernel\n",
+				__func__, devtools[i].fmt.c_str());
+			continue;
+		}
 		if (!access(devtools[i].fsck.c_str(), X_OK))
 			devtools[i].fsck_supported = true;
 		if (!access(devtools[i].mkfs.c_str(), X_OK))
@@ -451,7 +497,7 @@ void CHDDMenuHandler::setRecordPath(std::string &dev)
 		return;
 	}
 	/* don't annoy if the recordingdir is a symlink pointing to the 'right' location */
-	std::string readl = backtick("readlink -f " + g_settings.network_nfs_recordingdir);
+	std::string readl = readlink(g_settings.network_nfs_recordingdir.c_str());
 	readl = trim(readl);
 	if (newpath.compare(readl) == 0) {
 		printf("CHDDMenuHandler::%s: recordingdir is a symlink to %s\n",
@@ -699,8 +745,13 @@ bool CHDDMenuHandler::scanDevices()
 		return false;
 	}
 
+	int drive_mask = 0xfff0;
 	if (stat("/", &s) != -1)
-		root_dev = (s.st_dev & 0x0fff0); /* hda = 0x0300, hdb = 0x0340 */
+	{
+		if ((s.st_dev & drive_mask) == 0x0300) /* hda, hdb,... has max 63 partitions */
+			drive_mask = 0xffc0; /* hda: 0x0300, hdb: 0x0340, sda: 0x0800, sdb: 0x0810 */
+		root_dev = (s.st_dev & drive_mask);
+	}
 	printf("HDD: root_dev: 0x%04x\n", root_dev);
 
 	for(int i = 0; i < n;i++) {
@@ -709,6 +760,7 @@ bool CHDDMenuHandler::scanDevices()
 		char model[128] = { 0 };
 		int64_t bytes = 0;
 		int64_t megabytes;
+		bool oldkernel = false;
 		bool isroot = false;
 
 		printf("HDD: checking /sys/block/%s\n", namelist[i]->d_name);
@@ -720,7 +772,7 @@ bool CHDDMenuHandler::scanDevices()
 
 			int ret = fstat(fd, &s);
 			if (ret != -1) {
-				if ((int)(s.st_rdev & 0x0fff0) == root_dev) {
+				if ((int)(s.st_rdev & drive_mask) == root_dev) {
 					isroot = true;
 					/* dev_t is different sized on different architectures :-( */
 					printf("-> root device is on this disk 0x%04x, skipping\n", (int)s.st_rdev);
@@ -745,9 +797,16 @@ bool CHDDMenuHandler::scanDevices()
 		if (f) {
 			fscanf(f, "%s", vendor);
 			fclose(f);
+		} else {
+			oldkernel = true;
+			strcpy(vendor, "");
 		}
 
-		snprintf(str, sizeof(str), "/sys/block/%s/device/model", namelist[i]->d_name);
+		/* the Tripledragon only has kernel 2.6.12 available.... :-( */
+		if (oldkernel)
+			snprintf(str, sizeof(str), "/proc/ide/%s/model", namelist[i]->d_name);
+		else
+			snprintf(str, sizeof(str), "/sys/block/%s/device/model", namelist[i]->d_name);
 		f = fopen(str, "r");
 		if(!f) {
 			printf("Cant open %s\n", str);
@@ -796,6 +855,7 @@ int CHDDMenuHandler::doMenu()
 	show_menu = false;
 	in_menu = true;
 
+	check_kernel_fs();
 	check_dev_tools();
 
 _show_menu:
@@ -843,9 +903,11 @@ _show_menu:
 	}
 
 	if (!hdd_list.empty()) {
-		struct stat rec_st;
+		struct stat rec_st, root_st, dev_st;
 		memset(&rec_st, 0, sizeof(rec_st));
+		memset(&root_st, 0, sizeof(root_st));
 		stat(g_settings.network_nfs_recordingdir.c_str(), &rec_st);
+		stat("/", &root_st);
 
 		sort(hdd_list.begin(), hdd_list.end(), cmp_hdd_by_name());
 		int shortcut = 1;
@@ -859,6 +921,13 @@ _show_menu:
 			}
 			std::string key = "m" + it->devname;
 			bool enabled = !rec_icon || !CNeutrinoApp::getInstance()->recordingstatus;
+			/* do not allow to unmount the rootfs, and skip filesystems without kernel support */
+			memset(&dev_st, 0, sizeof(dev_st));
+			if (stat(("/dev/" + it->devname).c_str(), &dev_st) != -1
+			    && dev_st.st_rdev == root_st.st_dev)
+				enabled = false;
+			else if (kernel_fs_list.find(it->fmt) == kernel_fs_list.end())
+				enabled = false;
 			it->cmf = new CMenuForwarder(it->desc, enabled, g_Locale->getText(it->mounted ? LOCALE_HDD_UMOUNT : LOCALE_HDD_MOUNT) , this,
 					key.c_str(), CRCInput::convertDigitToKey(shortcut++), NULL, rec_icon);
 			hddmenu->addItem(it->cmf);
@@ -1017,6 +1086,7 @@ int CHDDMenuHandler::formatDevice(std::string dev)
 	int res;
 	FILE * f;
 	CProgressWindow * progress;
+	std::string fdisk, sfdisk, tune2fs;
 
 	printf("CHDDMenuHandler::formatDevice: dev %s hdd_fs %d\n", dev.c_str(), g_settings.hdd_fs);
 
@@ -1056,27 +1126,34 @@ int CHDDMenuHandler::formatDevice(std::string dev)
 		goto _return;
 	}
 
-#ifdef ASSUME_MDEV
-	creat("/tmp/.nomdevmount", 00660);
-#else
+#ifndef ASSUME_MDEV
 	f = fopen("/proc/sys/kernel/hotplug", "w");
 	if(f) {
 		fprintf(f, "none\n");
 		fclose(f);
 	}
 #endif
+	creat("/tmp/.nomdevmount", 00660);
 
 	progress = new CProgressWindow();
 	progress->setTitle(LOCALE_HDD_FORMAT);
 	progress->exec(NULL,"");
 	progress->showGlobalStatus(0);
 
-	if (access("/sbin/sfdisk", X_OK) == 0) {
-		snprintf(cmd, sizeof(cmd), "/sbin/sfdisk -f -uM %s", devname.c_str());
+	fdisk   = find_executable("fdisk");
+	sfdisk  = find_executable("sfdisk");
+	tune2fs = find_executable("tune2fs");
+	if (! sfdisk.empty()) {
+		snprintf(cmd, sizeof(cmd), "%s -f -uM %s", sfdisk.c_str(), devname.c_str());
 		strcpy(cmd2, "0,\n;\n;\n;\ny\n");
+	} else if (! fdisk.empty()) {
+		snprintf(cmd, sizeof(cmd), "%s -u %s", fdisk.c_str(), devname.c_str());
+		strcpy(cmd2, "o\nn\np\n1\n2048\n\nw\n");
 	} else {
-		snprintf(cmd, sizeof(cmd), "/sbin/fdisk %s", devname.c_str());
-		strcpy(cmd2, "o\nn\np\n1\n\n\nw\n");
+		/* cannot do anything */
+		fprintf(stderr, "CHDDFmtExec: neither fdisk nor sfdisk found in $PATH :-(\n");
+		showError(LOCALE_HDD_FORMAT_FAILED);
+		goto _remount;
 	}
 	progress->showStatusMessageUTF(cmd);
 
@@ -1186,20 +1263,19 @@ int CHDDMenuHandler::formatDevice(std::string dev)
 	}
 	sleep(2);
 
-	if (devtool->fmt.substr(0, 3) == "ext") {
+	if (devtool->fmt.substr(0, 3) == "ext" && ! tune2fs.empty()) {
 		std::string d = "/dev/" + devpart;
-		printf("CHDDMenuHandler::formatDevice: executing %s %s\n","/sbin/tune2fs -r 0 -c 0 -i 0", d.c_str());
-		my_system(8, "/sbin/tune2fs", "-r", "0", "-c", "0", "-i", "0", d.c_str());
+		printf("CHDDMenuHandler::formatDevice: executing %s %s %s\n", tune2fs.c_str(), "-r 0 -c 0 -i 0", d.c_str());
+		my_system(8, tune2fs.c_str(), "-r", "0", "-c", "0", "-i", "0", d.c_str());
 	}
 	show_menu = true;
 
 _remount:
+	unlink("/tmp/.nomdevmount");
 	progress->hide();
 	delete progress;
 
-#ifdef ASSUME_MDEV
-	unlink("/tmp/.nomdevmount");
-#else
+#ifndef ASSUME_MDEV
 	f = fopen("/proc/sys/kernel/hotplug", "w");
 	if(f) {
 		fprintf(f, "/sbin/hotplug\n");
@@ -1238,6 +1314,39 @@ _remount:
 			snprintf(cmd, sizeof(cmd), "%s/plugins", dst.c_str());
 			safe_mkdir(cmd);
 			// sync();
+#if HAVE_TRIPLEDRAGON
+		/* on the tripledragon, we mount via fstab, so we need to add an
+		   fstab entry for dst */
+		FILE *g;
+		char *line = NULL;
+		unlink("/etc/fstab.new");
+		g = fopen("/etc/fstab.new", "w");
+		f = fopen("/etc/fstab", "r");
+		if (!g)
+			perror("open /etc/fstab.new");
+		else {
+			if (f) {
+				int ret;
+				while (true) {
+					size_t dummy;
+					ret = getline(&line, &dummy, f);
+					if (ret < 0)
+						break;
+					/* remove lines that start with the same disk we formatted
+					   devname is /dev/xda" */
+					if (strncmp(line, devname.c_str(), devname.length()) != 0)
+						fprintf(g, "%s", line);
+				}
+				free(line);
+				fclose(f);
+			}
+			/* now add our new entry */
+			fprintf(g, "%s %s auto defaults 0 0\n", partname.c_str(), dst.c_str());
+			fclose(g);
+			rename("/etc/fstab", "/etc/fstab.old");
+			rename("/etc/fstab.new", "/etc/fstab");
+		}
+#endif
 		}
 	}
 _return:
